+++ integration/src/tests/prepare_address.rs
@@ -1,9 +1,9 @@
 use super::IntegrationTest;
-use crate::utils::prepare_address_with_ckb_capacity;
+use crate::utils::address::prepare_address_with_ckb_capacity;
 
 fn test_prepare_address() {
-    let address = prepare_address_with_ckb_capacity(10000000000).expect("prepare ckb");
-    println!("address: {:?}", address.to_string());
+    let _address_with_200_ckb =
+        prepare_address_with_ckb_capacity(200_0000_0000).expect("prepare ckb");
 }
 
 inventory::submit!(IntegrationTest {
